Refactor: [M3-6280]: SecretTokenDialog
