+++ packages/manager/src/features/Profile/APITokens/APITokenTable.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import AddNewLink from 'src/components/AddNewLink';
 import Box from 'src/components/core/Box';
 import Grid from '@mui/material/Unstable_Grid2';
-import SecretTokenDialog from 'src/features/Profile/SecretTokenDialog';
 import TableRowEmptyState from 'src/components/TableRowEmptyState';
 import TableRowError from 'src/components/TableRowError';
 import Typography from 'src/components/core/Typography';
@@ -13,6 +12,7 @@ import { EditAPITokenDrawer } from './EditAPITokenDrawer';
 import { isWayInTheFuture } from './utils';
 import { PaginationFooter } from 'src/components/PaginationFooter/PaginationFooter';
 import { RevokeTokenDialog } from './RevokeTokenDialog';
+import { SecretTokenDialog } from 'src/features/Profile/SecretTokenDialog/SecretTokenDialog';
 import { StyledTableSortCell } from 'src/components/TableSortCell/StyledTableSortCell';
 import { Table } from 'src/components/Table';
 import { TableBody } from 'src/components/TableBody';
