--- packages/manager/src/features/Profile/SecretTokenDialog/index.tsx
@@ -1,2 +1,0 @@
-import SecretTokenDialog from './SecretTokenDialog';
-export default SecretTokenDialog;
